Cloud Functions/Connection to big query/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-bigquery